Moved reaction functions to a separated class
